Add files via upload
In order to optimize production costs, the metallurgical plant decided to reduce electricity consumption at the stage of steel processing. Steel is processed in a metal ladle. Molten steel is poured into a ladle and heated to the desired temperature with graphite electrodes. 

The goal was to build a ML model that predicts the temperature. The model will help reduce energy consumption at the stage of steel processing, in other words, optimize production costs.