Add files via upload
We were given the task of identifying the patterns that determine the success of the game in order to bet on a potentially popular product and plan an advertising campaign for 2017. 
Historical game sales data, user and expert ratings, genres and platforms (such as Xbox or PlayStation) are available from open sources. Data presented up to 2016.