Add files via upload
Service for the sale of used cars is developing an application to attract new customers. In it, customers can quickly find out the market value of their car. At our disposal are historical data: technical specifications, equipment and prices of cars. We need to build a model to determine the cost.

It is important for the client:

    quality of prediction;
    model prediction speed;
    model training speed.